+++ lua/rldx/init.lua
@@ -9,7 +9,7 @@ function M.reset()
 	require("rldx").setup()
 end
 
-M.VERSION = "0.1.0"
+M.VERSION = "0.2.0"
 
 M.contacts = {}
 
@@ -65,21 +65,46 @@ function M.setup(options)
 		key = sett.session.encryption_key,
 	}
 
-	-- Load Contacts Database
-	M.contacts, err = crud.load_contacts(
-		sett.options.filename,
-		true,
-		enc_opts
-	)
+	-- Load the catalog
+	function M.rldx_load_cmd()
+		M.contacts, err = crud.load_contacts(
+			sett.options.filename,
+			true,
+			enc_opts
+		)
+		if err ~= nil then
+			vim.notify("RLDX failed to load contacts", "error")
+			return
+		end
+	end
+	M.rldx_load_cmd()
 
 	-- Register the Source with nvim-cmp
 	cmp.register_source("cmp_rolodex", M.source.new())
+
+	vim.api.nvim_create_user_command(
+		"RldxLoad",
+		M.rldx_load_cmd,
+		{ nargs = 0 }
+	)
+
+	vim.api.nvim_create_user_command(
+		"RldxSave",
+		M.rldx_save_cmd,
+		{ nargs = 0 }
+	)
 
 	-- Register the Add Contact command
 	vim.api.nvim_create_user_command(
 		"RldxAdd", 
 		M.rldx_add_cmd,
-		{ nargs = 1 }
+		{ nargs = 0 }
+	)
+
+	vim.api.nvim_create_user_command(
+		"RldxDelete",
+		M.rldx_delete_cmd,
+		{ nargs = 0 }
 	)
 
 	vim.api.nvim_create_user_command(
@@ -97,9 +122,78 @@ function M.rldx_list_cmd(opts)
 	vim.notify(vim.inspect(M.contacts))
 end
 
+function M.rldx_delete_cmd(opts)
+	local name = vim.fn.input('[DELETE] Enter Name: ')
+	vim.cmd('redraw')
+	vim.cmd('echo ""')
+
+	if name == nil or name == "" then
+		vim.notify("RLDX tried to delete an invalid name", "error")
+		return
+	end
+
+	local filtered_contacts = {}
+	for _, contact in ipairs(M.contacts) do
+		if contact.label ~= name then
+			table.insert(filtered_contacts, contact)
+		end
+	end
+	M.contacts = filtered_contacts
+
+
+	enc_opts = {
+		encryption = sett.options.encryption,
+		key = sett.session.encryption_key,
+		hash_salt_len = sett.options.hash_salt_length,
+	}
+
+	ok, err = crud.save_contacts(
+		sett.options.filename,
+		algos.copy_table(M.contacts),
+		sett.options.schema_ver,
+		enc_opts
+	)
+
+	if ok == true then
+		vim.notify("RLDX deleted '" .. name .. "' from Catalog")
+	else
+		vim.notify("RLDX failed to delete contact from Catalog")
+		return
+	end
+end
+
+function M.rldx_save_cmd(opts)
+	enc_opts = {
+		encryption = sett.options.encryption,
+		key = sett.session.encryption_key,
+		hash_salt_len = sett.options.hash_salt_length,
+	}
+
+	ok, err = crud.save_contacts(
+		sett.options.filename,
+		algos.copy_table(M.contacts),
+		sett.options.schema_ver,
+		enc_opts
+	)
+
+	if ok == true then
+		vim.notify("RLDX saved Catalog")
+	else
+		vim.notify("RLDX failed to save Catalog")
+		return
+	end
+end
+
 -- Add a contact to catalog
 function M.rldx_add_cmd(opts)
-	local name = opts.args
+	local name = vim.fn.input('[ADD] Enter Name: ')
+	vim.cmd('redraw')
+	vim.cmd('echo ""')
+
+	if name == nil or name == "" then
+		vim.notify("RLDX tried to add an invalid name", "error")
+		return
+	end
 
 	table.insert(
 		M.contacts, 
@@ -112,6 +206,7 @@ function M.rldx_add_cmd(opts)
 	enc_opts = {
 		encryption = sett.options.encryption,
 		key = sett.session.encryption_key,
+		hash_salt_len = sett.options.hash_salt_length,
 	}
 
 	ok, err = crud.save_contacts(
@@ -122,9 +217,9 @@ function M.rldx_add_cmd(opts)
 	)
 
 	if ok == true then
-		vim.notify("Added '" .. name .. "' to Catalog")
+		vim.notify("RLDX added '" .. name .. "' to Catalog")
 	else
-		vim.notify("Failed to add contact to Catalog")
+		vim.notify("RLDX failed to add contact to Catalog")
 		return
 	end
 end