+++ lua/rldx/utils/crud.lua
@@ -3,14 +3,17 @@ local algos = require("rldx.utils.algos")
 
 local v0_0_2_mw = require("rldx.utils.v0_0_2.middleware")
 local v0_1_0_mw = require("rldx.utils.v0_1_0.middleware")
+local v0_2_0_mw = require("rldx.utils.v0_2_0.middleware")
 
 local M = {}
 
 function M.save_contacts(filepath, catalog, ver, opts)
 	if ver == "0.0.2" then
 		catalog = v0_0_2_mw.from_completions(catalog, opts)
-	elseif (ver == "0.1.0") or (ver == "latest") then
+	elseif (ver == "0.1.0") then
 		catalog = v0_1_0_mw.from_completions(catalog, opts)
+	elseif (ver == "0.2.0") or (ver == "latest") then
+		catalog = v0_2_0_mw.from_completions(catalog, opts)
 	elseif ver == nil then
 		vim.notify("invlaid schema version: nil", "error")
 		return false
@@ -55,6 +58,8 @@ function M.load_contacts(filepath, create, opts)
 		catalog = v0_0_2_mw.to_completions(catalog, opts)
 	elseif ver == "0.1.0" then
 		catalog = v0_1_0_mw.to_completions(catalog, opts)
+	elseif ver == "0.2.0" or (ver == "latest") then
+		catalog = v0_2_0_mw.to_completions(catalog, opts)
 	elseif ver == nil then
 		vim.notify("invlaid schema version: nil", "error")
 		return {}